pyproject.toml
[tool.poetry]
name = "acme"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.5"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.8.0"
black = "^22.12.0"


[tool.poetry.group.test.dependencies]
faker = "^16.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
